test _execute_job
